implemented change to theme and using in a themeToggler component
